+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 master
 ~~~~~~
 1) [FEATURE] Extend sys_fie_reference export in DCE with standard link field.
+2) [FEATURE] Add new backend and ext icon.
 
 0.1.0
 ~~~~~~