+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ master
 ~~~~~~
 1) [FEATURE] Extend sys_fie_reference export in DCE with standard link field.
 2) [FEATURE] Add new backend and ext icon.
+3) [FEATURE] Add Hugo build service and cli build task.
 
 0.1.0
 ~~~~~~